NovskyProject/requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
markdownx
